integrate postgres db and migrations using sqlx (#16)
* add backend repository

* try using the new sqlx connection

* renaming things

* add migration process

* remove call

* wip

* remove comment

* remove old structs

* forward postgres service to local

* add models

* add deployment status enum

* update protobuf version

* update models

* make hash and name unique

* add more realistic models

* update names

* fix fks

* fix migration order

---------

+++ internal/repository/beam_postgres.go
@@ -425,7 +425,7 @@ func (c *PostgresBeamRepository) GetTotalUsageOfUserMs(identity types.Identity) 
 	var identityConditions []string
 	args := make([]interface{}, 0)
 
-	err = c.client.Where(&types.BeamApp{IdentityId: identity.ID}).Find(&beamApps).Error
+	err = c.client.Where(&types.BeamApp{IdentityId: uint(identity.ID)}).Find(&beamApps).Error
 	if err != nil {
 		return 0, err
 	}
